More index tweaking/optimization and small thread hiding fix.
Most feature that need to be called on each index refresh don't need
to be called on every single page build, but only after building the
threads. ThreadHiding needs to be fired on each page build to ensure
that the stub summaries fit with the number of replies, so it has to
wait for the last replies to be added. This also fixes the bug where
ThreadHiding was creating one extra stub on each IndexRefresh event.